Display Aislings
Added initial code to the SeverMessage 0x34 when the packet is received to display an aisling. Aisling display data is stored in an instance of the Player class and a call is made to a new function in client.cs called DisplayAisling which displays an aisling from Player object data.

Another helper class was made in client.cs to Click an object which is called in the ServerMessage 0x34 method