Doors are done
In server.cs, added code to handle when the door packet is sent to a player (i.e., whenever a door becomes visible on a map, or a door is opened/closed, or a player is loaded onto a map with doors).

In client.cs, added a hashset to store door data as a Location (MapID, X, Y).

Added override methods to the Door.cs class to correctly calculate hashes for the hashset so that doors of the same Location were not being added continuously.