Major updates to Items/Spells
Added packet handling methods in server.cs to AddItemsToPane and AddSpellsToPane.
Required creation of Inventory and Spellbook classes to store items and spells that were added to each pane.
Added appropriate variables to Client.cs for Items and Spells

Created methods in client.cs to laod melee weapons, staves, and bows. Will need to update for new items in the game.