Major updates to process ServerMessages
The packet method that handles server messages (orange bar, etc...) was update to handle the various types of messages that come through. Most of them have been accounted for, and the majority of the working being done under the ActiveMessage type.

A new class called ActiveMessageHandler was created to process the large swaths of code related to the many "orange bar" messages that come through the game. A good chunk of messages were accounted for in the handler class but will need to be update to account for all of them.

A new method was created in Client.cs called UpdateCurseTargets which tries to update the status of the curse on each target when the orange message "another spell afflicts thee..." comes up in the orange message bar.

Server.cs has a new method called ProcessEffects, but that will be integrated into the ActiveMessageHandler as well.