Point, CONSTANTS, looting
Added methods for Point to take int parameters

Added a loot method to Bot.cs that checks for lootable objects within 4 spaces of the player. Added accompanying packetmethods for looting and dropping in Client.cs. Also created a method to getlootableobjects(int distance) away from player. It should probably just be default 4.

Updates to allypage to only display usable spells.

Created task threads for looting and a bot loop. Need to check what is going on with the threads because they only run once and we want them to run continuously.