Strangers
Updated code to update the stranger list and last seen aislings.

In ClientTab.cs we get nearby players and add them to the appropriate binding list if they are not allies. We also add them to a dictionary of last seen aisling. We also had to check if the client has used "charge-like" abilities because we dont want to charge past another player and not display them correctly and/or the charge skill gets buggy when updating the list. We use a separate method UpdateStrangerListAfterCharge that waits a couple of seconds before calling the updatestrangerlist.