NearbAlly updates
Added a NearbyAlly control that loads aislings nearby on screen that can be added as targeted aislings to cast buffs on etc.

Added methods to load Khan directories and load .pal files (palletes) to load character sprite data that represents what the player is currently wearing in game. They are displayed on the Nearby Aislings tab control. Currently implemented are aisling bodies and overcoats. Need to adjust placement of the files to overlap and appear "seamless". Need to implement other methods to load remaining palette files.