Rangers in world list
Added packet methods to server world list to display rangers in red, and add functionality to DC clients when ranger is found on list.

Method added to Server.cs called Automatic Actions and added to the server thread upon initialization. When the start button is pressed and the Ranger/GM checkbox is checked, it will constantly send the packet to request the world list, which in turn will trigger the Worldlist packet and disconnect clients when a ranger is found in the list.
-other neat things, auto buy komadiums
-update inventory (not implemented)
-update client status (name tbd, not implemented)