Spell animation updates
Added some updates to the animation packet..

todo: finish out the animations we want to track.
todo: decide if we want to try and remove current spells from _creatureToSpellList[0] in this method or if it will mess up the same kind of call in activemessagehandler "you cast"... like we remove the correct spell and then we remove another because we're calling it twice.