Packet handler updates
Added:
ConfirmCreatureWalk logic

Fixed:
PacketLog was previously displaying hex values of encrypted packets which is not useful in and of itself.
In the clientloop method of client.cs, apparently the packets passed to the clienttab method logpackets() were being modified before processing (potentially because needed to invoke the ui thread). To fix this I added copies of clientpacket and server packet to pass, instead of passing the actual client/server packets.