Major Concurrency changes
There was an issue where player object hashes were changing on refresh. An update call to display aislings was sending the server packet to DisplayAislings, and the Player constructor was being called repeatedly and overwriting player objects inside the server packet handler.

A new method was created called GetOrCreatePlayer in Server.cs to grab the player from WorldObjects. If it is an exisiting object it will assign the value and return it back to the calling method. If it did not exist, it is created on and then passed back to the calling method.

All dictionaries that include Player types were changed to ConcurrentDictionaries and now use AddOrUpdate to add values and TryRemove to remove values.

Methods to clear dictionaries with Player types were removed from the MapInfo server packet handler and moved to the map change server packet handler.

Various fixes