Cooldown changes
Added methods to both Skillbook.cs and Spellbook.cs classes to update the values of cooldown and ticks. These methods are called in the Server Packet Handler 0x3F_Cooldown.

The previous implementation was not properly handling cooldown and ticks for Spells and Skills, and was assigning them to different objects (different hashes). This approach ensures that we are updating the correct skill and spell objects.

Updated the CanUse Property for both Skills and Spells. The new implementation is a more straightforward approach that tests whether the current time is > Ticks (expressed in time).

Tested and working