Fuck pathfinding again
Pathfinding seems to be working.
Implemented a new class PriorityQueue because .net framework does not have one natively.

Reworked TryWalkToLocation method to account for pathfiding nuances and _walkSpeed throttle.
Had to implement a _walkSpeed throttle on the Walk method as well... probably some multithreaded bullshit.
Consider adding a class to handle Walk timing.... but for now it's working....