Minor pathfind modifications
When the display aisling server packet handler is called, we needed to add the ID of the aisling to the creaturehashset. This allows us to properly call GetWorldObjects() and add those aislings to the obstacle list in pathfinder.

Also changed the nomenclature of Pathfinder  from ignoreObstacles to avoidObstacles to properly account for obstacle avoidance when finding paths.

Added another call to intializePathNodes inside of FindPath to dynamically update new obstacles. Might have a higher overhead though...