Route and Pathfinding changes
Refactored some of Routefinder.cs
Added a check to Pathfinder.cs to check if the pathnode is an Exit on the map, because doorways were being evaluated as walls. So the new logic checks nodes Open if it is not a wall or is an exit. So we can path to buildings.

Changed doors from a hashset to a concurrentdictionary and methods associated. Added a line to the Door server message handler to set the value of door to the current value in the dictionary. This was important because we weren't updated the door state in the door dictionary.