Display menu server packet handler
Added implementation for when the server displays a menu to the player.
Pursuit IDs are recorded and displayed on the info tab.
If the player is withdrawing items from the bank, it will store a list of items to a text file in base directiory/inventory/playername/bank.txt

Because we implemented storage of pursuit IDs, routefinding between temuair and medenia server maps via loures harbor and noam carpet merchant are working now.