Walker updates.. i.e., Nobis maze is fucking stupid
Fleshed out the Walker loop by implementing several new methods to handle walking based on the walkMapCombo box.

Need to write the following methods:
FollowWalking
WayPointWalking
SWLure

Also need to check the implementation of Client.SetLocationByMapID as it sets a location with a map ID and coordiantes of 0,0. Right now it tries to walk to 0, 0. Need to check TryWalkToLocation and check for a 0, 0 location and break from the Method if Map IDs are equal. That would drop us at the warp inside the map location and not try to walk any further.