Creature state persistence
This fucking issue again...

Implemented a new class CreatureStateHelper that is called anytime the state of a tracked spell should be updated, and updates it across all connected clients.

This implementation completely replaces methods client.UpdateCurseTragets and client.UpdateFasTargets.

The CreatureStateHelper class receives a generic object of type CreatureState, an enum that was added to hold the state. It iterates through all connected clients and updates the State of the creature ID in the WorldObjects dictionary. In this way, we have a scalable and consistent way to maintain the state of tracked spells for all clients.

Modified several places in the codebase where this should happen.

Tested with curse and fas.

Todo: Modify the creature class to remove existing variables and reimplement them using CreatureState and CreatureStateHelper.