State persistence... final boss edition
The pervious iteration of CreatureStateHlper was working well.. when both clients were connected and both clients had a copy of the creature object in WorldObjects.

This means that whenever the lead client was walking forward and casting, it was trying to update states across all clients copies of WorldObjects, but for clients not in view of the creature this did not happen.

To solve this a cache was creatued for CreatureStates in creaturestate helper.
If the object is not found in a client's worldobjects dictionary, it is stored in the _pendingUpdates dictionary inside of the creatureStateHelper class via CachePendingUpdate.

When the 0x07 packet do display a monster is called, and when the subsequent method to GetorCreateCreature is called, we call CreatureStateHelper.ApplyCachedUpdates to make sure that all pending updates are applied to creatures when they become visible. This allows us to maintain state visibility for all creatures across all clients.

**Consider adding a method to remove "timedout" pending updates so as not to be a resource hog