Stuck fix
Modified methods to achieve a fix for being "stuck" or when another player is on top of you. Particularly relevant when game rangers port to you, or if another player is trying to murder you while hunting.

The bot will handle the stuck message, get a list of the surrounding tiles and move to the first one that is available.