Debugging packet handling for Tavaly bashing
Still haven't found out what is exactly causing the DCing.

Changed from assails using UseSkill in quick succession to sending a spacebar packet (did not fix)

Tried dynamically resizing packetbuffers (did not work at all)

Tried introducing a skill queue for UseSkills but the skills were not being used as often as we want for hunting, even with a 100ms delay.

Still up in the air..