Renaming, Extensions, and Profile saving
Updated list of items that saves and loads on profiles.

Added a Jobject extension class in anticipation for managing autoascending data from json files.

Added a "farthest first" option for spelling, so that when running 2 characters that are using the same spell (2 arders) so the leader can cast on the farthest mobs and the follower can cast on nearest mobs. Good for not overlapping spell attempts for same spells.