Hidden player names
Fixed an issue where names were not always updating on hidden players, or the profile was getting stuck on the screen after clicking the player to get their name.

Ran into an issue where this was repeatedly looping when multiple clients were hidden. Chose to skip the logic to display names above heads for connected clients.